Refresh SyncShell overrides after settings changes
